requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
